refactor: separate DarkModeContext for Fast Refresh compliance
- Move context creation to DarkModeContext.types.ts
- Keep only DarkModeProvider component in DarkModeContext.tsx
- Update imports in useDarkModeContext hook
- Fix ESLint react-refresh/only-export-components warning
- Use type-only import for ReactNode to comply with verbatimModuleSyntax
