+++ web/src/contexts/DarkModeContext.tsx
@@ -1,12 +1,5 @@
-import { createContext, useState, useEffect, ReactNode } from 'react';
-
-export interface DarkModeContextType {
-  isDarkMode: boolean;
-  toggleDarkMode: () => void;
-  setDarkMode: (value: boolean) => void;
-}
-
-export const DarkModeContext = createContext<DarkModeContextType | undefined>(undefined);
+import { useState, useEffect, type ReactNode } from 'react';
+import { DarkModeContext } from './DarkModeContext.types';
 
 const DARK_MODE_KEY = 'english-playlist-dark-mode';
