+++ web/src/contexts/DarkModeContext.types.ts
@@ -1,0 +1,9 @@
+import { createContext } from 'react';
+
+export interface DarkModeContextType {
+  isDarkMode: boolean;
+  toggleDarkMode: () => void;
+  setDarkMode: (value: boolean) => void;
+}
+
+export const DarkModeContext = createContext<DarkModeContextType | undefined>(undefined);
